# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API documentation for use marketplace database.",
        "title": "Swagger Marketplace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6060",
    "basePath": "/api/v1",
    "paths": {
...
        "/products/": {
            "post": {
                "description": "post product data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Insert product data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            }
        },
        "/products/:id": {
            "get": {
                "description": "get product data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "put product data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Edit product data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete product data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Discard product data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.Product": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.ProductResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Product"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}